Refactoring of common types for network graphs
