+++ src/components/Visualizations/GraphPanes/types.ts
@@ -1,0 +1,7 @@
+import { IADSApiAuthorNetworkNodeKey, IADSApiPaperNetworkNodeKey } from '@api';
+
+export interface IView {
+  id: string;
+  label: string;
+  valueToUse: IADSApiPaperNetworkNodeKey | IADSApiAuthorNetworkNodeKey;
+}
