+++ src/components/Visualizations/Graphs/types.ts
@@ -1,0 +1,5 @@
+// extend the d3 Path element to cache the last angle
+// Used for transitioning for animated transition to new angle
+export interface ADSSVGPathElement extends SVGPathElement {
+  lastAngle: { x0: number; x1: number };
+}
